requirements.txt
redis[hiredis]
fastapi[standard]
uvicorn[standard]
requests-cache
jinja2
google-generativeai